+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 
 ### :bug: Bug Fix
 
+- `plugin-common`
+  - [#3280](https://github.com/wix/ricos/pull/3280) fix plugin toolbar position in rtl
 - `collapsible-list`
   - [#3297](https://github.com/wix/ricos/pull/3297) fix re-render answer content on collpase/expand